+++ .gitignore
@@ -2,11 +2,7 @@ node_modules
 coverage
 dist
 /.env
-docker-compose.override.yml
 __pycache__
-packages/code-du-travail-data/dump.json
-packages/code-du-travail-data/dataset/suggestions.txt
-package-lock.json
 legacy
 *.log
 data/
@@ -15,3 +11,5 @@ large-*
 cache.json
 data.csv
 suggestions.txt
+bin
+lib
