+++ Dockerfile
@@ -7,11 +7,11 @@ COPY ./tsconfig.json tsconfig.json
 COPY ./src/ src/
 
 RUN yarn install
-RUN yarn build
+RUN yarn build:bin
 
 FROM node:14.16-alpine3.10
 WORKDIR /app
-COPY --from=builder /app/dist/ .
+COPY --from=builder /app/bin .
 
 
 CMD ["sh", "-c", "node index.js ${MONOLOG_ACTION}"]
