--- index.ts
@@ -1,3 +1,0 @@
-import { LogQueries } from "./src/logQueries";
-
-export { LogQueries };
